FIX: When 7-zip is embedded in some archive it is automatically opened even when there are other files beside it
The problem is that 7zr.exe seems to automatically scan supported archives for a .7z archive inside it and it returns the embedded archive if its found as if it was at top level.

The workaround is to use 7za.exe which supports all archive types and check explicitly what is the type of the archive.