FIX: Crash during startup when the default application's language has been left on first run and then explicitly set later.
The problem was that during saving of settings the config file was opened in a mode causing the data to be overwritten, instead of file being truncated and then recreated. On the first run the application has been writing the language as "OSLang" and when later changed the value changed to "en-US" (for example), which is 1 char shorter thus causing the file to contain a last character from its previous state in addition to current state.